+++ R/ModStatR-package.R
@@ -1,10 +1,13 @@
-#' @title ModStatR
+#' @keywords internal
+#' @name ModStatR-package
+#' @aliases ModStatR-package ModStatR NULL
 #'
-#' @description Motivation: Package compagnon du livre Modélisation statistique par la pratique avec R. Il contient les codes des chapitres du livre ainsi que les solutions des exercices mais aussi d'autres compléments à découvrir.
-#'
-#' @docType package
-#' @name ModStatR
 #' @references \emph{Modélisation statistique par la pratique avec R}, Frédéric Bertrand, Emmanuelle Claeys, Myriam Maumy-Bertrand, 2019, ISBN:9782100793525, Dunod, Paris, \url{https://www.dunod.com/sciences-techniques/modelisation-statistique-par-pratique-avec-r-cours-et-exercices-corriges}, \url{https://github.com/fbertran/ModStatR/} et \url{https://fbertran.github.io/ModStatR/}
+#' 
+#' @examples
+#' set.seed(314)
+"_PACKAGE"
+
 #' @importFrom graphics abline lines par plot points polygon rect
 #' @importFrom stats qt cor cor.test integrate na.omit
 #' @importFrom boot boot.ci
@@ -13,6 +16,4 @@
 #' @importFrom jmuOutlier perm.cor.test
 #' @import ellipse
 #' 
-#' @examples
-#' set.seed(314)
 NULL