+++ R/datasets.R
@@ -360,7 +360,7 @@
 
 #' @title Sida du chat
 #' @description Taux de leucocytes
-#' @format Un objet \code{data.frame} avec 33 lignes et 3 variables :
+#' @format Un objet \code{data.frame} avec 32 lignes et 3 variables :
 #' \describe{
 #'   \item{\code{Sexe}}{Factor w/ 2 levels, Sexe de l'animal}
 #'   \item{\code{Jours}}{Integer vector, Nombre de jours après l'inoculation}